+++ src/extension.ts
@@ -1,13 +1,11 @@
 'use strict';
 
-import * as path from 'path';
-
-import { workspace, ExtensionContext, Hover, MarkdownString, ProviderResult } from 'vscode';
+import { ExtensionContext } from 'vscode';
 import * as vscode from 'vscode';
 import * as solargraph from 'solargraph-utils';
-import { LanguageClient, LanguageClientOptions, ServerOptions, TransportKind, Middleware, RequestType, MessageTransports, createClientSocketTransport, Disposable } from 'vscode-languageclient';
-import SolargraphDocumentProvider from './SolargraphDocumentProvider';
+import { LanguageClient, Disposable } from 'vscode-languageclient';
 import { makeLanguageClient } from './language-client';
+import SolargraphWebviewProvider from './SolargraphWebviewProvider';
 
 export function activate(context: ExtensionContext) {
 
@@ -23,92 +21,41 @@ export function activate(context: ExtensionContext) {
 	let solargraphConfiguration = new solargraph.Configuration();
 	applyConfiguration(solargraphConfiguration);
 	
-	let solargraphDocumentProvider = new SolargraphDocumentProvider();
-
 	let languageClient: LanguageClient;
 	let disposableClient: Disposable;
+	let webViewProvider: SolargraphWebviewProvider = new SolargraphWebviewProvider();
 
 	var startLanguageServer = function() {
 		languageClient = makeLanguageClient(solargraphConfiguration);
-		solargraphDocumentProvider.setLanguageClient(languageClient);
+		languageClient.onReady().then(() => {
+			languageClient.onNotification('$/solargraph/restart', (params) => {
+				console.log('I should restart!');
+			});
+		});
 		disposableClient = languageClient.start();
+		webViewProvider.setLanguageClient(languageClient);
 		context.subscriptions.push(disposableClient);
-	}
-
-	// https://css-tricks.com/snippets/javascript/get-url-variables/
-	var getQueryVariable = function (query, variable) {
-		var vars = query.split("&");
-		for (var i = 0; i < vars.length; i++) {
-			var pair = vars[i].split("=");
-			if (pair[0] == variable) { return pair[1]; }
-		}
 	}
 
 	var restartLanguageServer = function (): Promise<void> {
 		return new Promise((resolve) => {
-			/*if (languageClient) {
+			if (languageClient) {
 				languageClient.stop().then(() => {
 					disposableClient.dispose();
-					socketProvider.restart().then(() => {
-						languageClient = makeLanguageClient(socketProvider);
-						solargraphDocumentProvider.setLanguageClient(languageClient);
-						disposableClient = languageClient.start();
-						context.subscriptions.push(disposableClient);
-						resolve();
-					});
+					startLanguageServer();
+					resolve();
 				});
 			} else {
 				startLanguageServer();
 				resolve();
-			}*/
-			resolve();
-		});
-	}
-
-	function checkGemVersion() {
-		console.log('Checking gem version');
-		solargraph.verifyGemIsCurrent(solargraphConfiguration).then((result) => {
-			if (result) {
-				console.log('Solargraph gem version is current');
-			} else {
-				notifyGemUpdate();	
 			}
-		}).catch(() => {
-			console.log('An error occurred checking the Solargraph gem version.');
 		});
 	}
-	
-	function notifyGemUpdate() {
-		if (vscode.workspace.getConfiguration('solargraph').useBundler && vscode.workspace.rootPath) {
-			vscode.window.showInformationMessage('A new version of the Solargraph gem is available. Update your Gemfile to install it.');
-		} else {
-			vscode.window.showInformationMessage('A new version of the Solargraph gem is available. Run `gem update solargraph` to install it.', 'Update Now').then((item) => {
-				if (item == 'Update Now') {
-					solargraph.updateGem(solargraphConfiguration).then(() => {
-						restartLanguageServer().then(() => {
-							vscode.window.showInformationMessage('Successfully updated the Solargraph gem.');
-						});
-					}).catch(() => {
-						vscode.window.showErrorMessage('Failed to update the Solargraph gem.');
-					});
-				}
-			});
-		}
-	}
-
-	// Open command (used internally for browsing documentation pages)
-	var disposableOpen = vscode.commands.registerCommand('solargraph._openDocument', (uriString: string) => {
-		var uri = vscode.Uri.parse(uriString);
-		var label = (uri.path == '/search' ? 'Search for ' : '') + getQueryVariable(uri.query, "query");
-		vscode.commands.executeCommand('vscode.previewHtml', uri, vscode.ViewColumn.Two, label);
-	});
-	context.subscriptions.push(disposableOpen);
 
 	// Open URL command (used internally for browsing documentation pages)
 	var disposableOpenUrl = vscode.commands.registerCommand('solargraph._openDocumentUrl', (uriString: string) => {
 		var uri = vscode.Uri.parse(uriString);
-		var label = (uri.path == '/search' ? 'Search for ' : '') + getQueryVariable(uri.query, "query");
-		vscode.commands.executeCommand('vscode.previewHtml', uri, vscode.ViewColumn.Two, label);
+		webViewProvider.open(uri);
 	});
 	context.subscriptions.push(disposableOpenUrl);
 
@@ -126,7 +73,7 @@ export function activate(context: ExtensionContext) {
 		vscode.window.showInputBox({prompt: 'Search Ruby documentation:'}).then(val => {
 			if (val) {
 				var uri = 'solargraph:/search?query=' + encodeURIComponent(val);
-				vscode.commands.executeCommand('solargraph._openDocument', uri);
+				webViewProvider.open(vscode.Uri.parse(uri));
 			}
 		});
 	});
@@ -142,28 +89,35 @@ export function activate(context: ExtensionContext) {
 
 	// Check gem version command
 	var disposableCheckGemVersion = vscode.commands.registerCommand('solargraph.checkGemVersion', () => {
-		// languageClient.sendNotification('$/solargraph/checkGemVersion', { verbose: true });
-		solargraph.verifyGemIsCurrent(solargraphConfiguration).then((result) => {
-			if (result) {
-				vscode.window.showInformationMessage('The Solargraph gem is up to date.');
-			} else {
-				notifyGemUpdate();	
-			}
-		}).catch(() => {
-			console.log('An error occurred checking the Solargraph gem version.');
-		});
+		languageClient.sendNotification('$/solargraph/checkGemVersion', { verbose: true });
 	});
-	context.subscriptions.push(disposableSearch);
+	context.subscriptions.push(disposableCheckGemVersion);
 	
 	// Build gem documentation command
 	var disposableBuildGemDocs = vscode.commands.registerCommand('solargraph.buildGemDocs', () => {
-		languageClient.sendNotification('$/solargraph/documentGems', { rebuild: false });
+		let prepareStatus = vscode.window.setStatusBarMessage('Building new gem documentation...');
+		languageClient.sendRequest('$/solargraph/documentGems', { rebuild: false }).then((response) => {
+			prepareStatus.dispose();
+			if (response['status'] == 'ok') {
+				vscode.window.setStatusBarMessage('Gem documentation complete.', 3000);
+			} else {
+				vscode.window.setStatusBarMessage('An error occurred building gem documentation.', 3000);
+			}
+		});
 	});
 	context.subscriptions.push(disposableBuildGemDocs);
 
 	// Rebuild gems documentation command
 	var disposableRebuildAllGemDocs = vscode.commands.registerCommand('solargraph.rebuildAllGemDocs', () => {
-		languageClient.sendNotification('$/solargraph/documentGems', { rebuild: true });
+		let prepareStatus = vscode.window.setStatusBarMessage('Rebuilding all gem documentation...');
+		languageClient.sendRequest('$/solargraph/documentGems', { rebuild: true }).then((response) => {
+			prepareStatus.dispose();
+			if (response['status'] == 'ok') {
+				vscode.window.setStatusBarMessage('Gem documentation complete.', 3000);
+			} else {
+				vscode.window.setStatusBarMessage('An error occurred rebuilding gem documentation.', 3000);
+			}
+		});
 	});
 	context.subscriptions.push(disposableRebuildAllGemDocs);
 
@@ -182,28 +136,22 @@ export function activate(context: ExtensionContext) {
 	
 	// Solargraph download core command
 	var disposableSolargraphDownloadCore = vscode.commands.registerCommand('solargraph.downloadCore', () => {
-		var child = solargraph.commands.solargraphCommand(['download-core'], solargraphConfiguration);
-		child.on('exit', (code) => {
-			if (code == 0) {
-				vscode.window.showInformationMessage('Core documentation downloaded.');
-			} else {
-				vscode.window.showErrorMessage('Error downloading core documentation.');
-			}
-		});
-		// TODO: LSP version of downloadCore
-		// if (languageClient) {
-		// 	languageClient.sendNotification('$/solargraph/downloadCore');
-		// } else {
-		// 	vscode.window.showInformationMessage('Solargraph is still starting. Please try again in a moment.');
-		// }
+		if (languageClient) {
+			languageClient.sendNotification('$/solargraph/downloadCore');
+		} else {
+			vscode.window.showInformationMessage('Solargraph is still starting. Please try again in a moment.');
+		}
 	});
 	context.subscriptions.push(disposableSolargraphDownloadCore);
 
 	startLanguageServer();
-
-	vscode.workspace.registerTextDocumentContentProvider('solargraph', solargraphDocumentProvider);
+	languageClient.onReady().then(() => {
+		if (vscode.workspace.getConfiguration('solargraph').checkGemVersion) {
+			languageClient.sendNotification('$/solargraph/checkGemVersion', { verbose: false });
+		}
+	});
 }
 
 export function deactivate() {
-	// TODO Stop the language server?
+	// TODO: Any cleanup necessary?
 }